+++ android/NUnit-Appium-App-BrowserStack/NUnit-Appium-App-BrowserStack.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 25.0.1705.1
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NUnit-Appium-App-BrowserStack", "NUnit-Appium-App-BrowserStack\NUnit-Appium-App-BrowserStack.csproj", "{4F27E3E3-BB10-48D5-8ADC-3070FE32BA05}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{4F27E3E3-BB10-48D5-8ADC-3070FE32BA05}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4F27E3E3-BB10-48D5-8ADC-3070FE32BA05}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4F27E3E3-BB10-48D5-8ADC-3070FE32BA05}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4F27E3E3-BB10-48D5-8ADC-3070FE32BA05}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {0DE98670-62A6-4E0D-B5F7-B0609308F7DB}
+	EndGlobalSection
+EndGlobal
