--- android/appium_dotnet_driver_3_examples/App.config
@@ -1,56 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<configuration>
-    <configSections>
-        <sectionGroup name="capabilities">
-            <section name="first" type="System.Configuration.AppSettingsSection, System.Configuration, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a"/>
-            <section name="local" type="System.Configuration.AppSettingsSection, System.Configuration, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a"/>
-            <section name="parallel" type="System.Configuration.AppSettingsSection, System.Configuration, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a"/>
-        </sectionGroup>
-        <sectionGroup name="environments">
-            <section name="pixel-3" type="System.Configuration.AppSettingsSection, System.Configuration, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a"/>
-            <section name="galaxy-s10e" type="System.Configuration.AppSettingsSection, System.Configuration, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a"/>
-        </sectionGroup>
-    </configSections>
-
-    <appSettings>
-        <add key="user" value="<browserstack-user-name>"/>
-        <add key="key" value="<browserstack-access-key>"/>
-        <add key="server" value="hub.browserstack.com"/>
-    </appSettings>
-
-    <capabilities>
-        <first>
-            <add key="project" value="NUnit Android"/>
-            <add key="build" value="First NUnit Android Test"/>
-            <add key="name" value="first_test"/>
-            <add key="browserstack.debug" value="true"/>
-            <add key="app" value="bs://<app_hashed_id>"/>
-        </first>
-        <local>
-            <add key="project" value="NUnit Android"/>
-            <add key="build" value="Local NUnit Android Test"/>
-            <add key="name" value="local_test"/>
-            <add key="browserstack.debug" value="true"/>
-            <add key="browserstack.local" value="true"/>
-            <add key="app" value="bs://<app_hashed_id>"/>
-        </local>
-        <parallel>
-            <add key="project" value="NUnit Android"/>
-            <add key="build" value="Parallel NUnit Android Test"/>
-            <add key="name" value="parallel_test"/>
-            <add key="browserstack.debug" value="true"/>
-            <add key="app" value="bs://<app_hashed_id>"/>
-        </parallel>
-    </capabilities>
-
-    <environments>
-        <pixel-3>
-            <add key="device" value="Google Pixel 3"/>
-            <add key="os_version" value="9.0"/>
-        </pixel-3>
-        <galaxy-s10e>
-            <add key="device" value="Samsung Galaxy S10e"/>
-            <add key="os_version" value="9.0"/>
-        </galaxy-s10e>
-    </environments>
-</configuration>
