--- android/appium_dotnet_driver_3_examples/packages.config
@@ -1,12 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Appium.WebDriver" version="3.0.0.2" targetFramework="net461" />
-  <package id="BrowserStackLocal" version="1.4.0" targetFramework="net461" />
-  <package id="Castle.Core" version="4.3.1" targetFramework="net461" />
-  <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net461" />
-  <package id="NUnit" version="3.8.0" targetFramework="net461" />
-  <package id="NUnit3TestAdapter" version="3.13.0" targetFramework="net461" />
-  <package id="Selenium.Support" version="3.6.0" targetFramework="net461" />
-  <package id="Selenium.WebDriver" version="3.6.0" targetFramework="net461" />
-  <package id="System.Runtime.InteropServices.RuntimeInformation" version="4.0.0" targetFramework="net461" />
-</packages>
