--- android/appium_dotnet_driver_3_examples/parallel-test/ParallelTest.cs
@@ -1,32 +1,0 @@
-﻿using System;
-using System.Threading;
-using System.Collections.ObjectModel;
-using NUnit.Framework;
-using OpenQA.Selenium;
-using OpenQA.Selenium.Appium;
-using OpenQA.Selenium.Appium.Android;
-using OpenQA.Selenium.Support.UI;
-
-namespace android.parallel
-{
-  [TestFixture("parallel", "pixel-3")]
-  [TestFixture("parallel", "galaxy-s10e")]
-  [Parallelizable(ParallelScope.Fixtures)]
-  public class ParallelTest : BrowserStackNUnitTest
-  {
-    public ParallelTest(string profile, string device) : base(profile,device){ }
-
-    [Test]
-    public void searchWikipedia()
-    {
-      AndroidElement searchElement = (AndroidElement)new WebDriverWait(driver, TimeSpan.FromSeconds(30)).Until(ExpectedConditions.ElementToBeClickable(MobileBy.AccessibilityId("Search Wikipedia")));
-      searchElement.Click();
-      AndroidElement insertTextElement = (AndroidElement)new WebDriverWait(driver, TimeSpan.FromSeconds(30)).Until(ExpectedConditions.ElementToBeClickable(By.Id("org.wikipedia.alpha:id/search_src_text")));
-      insertTextElement.SendKeys("Browserstack");
-      Thread.Sleep(5000);
-
-      ReadOnlyCollection<AndroidElement> allProductsName = driver.FindElements(By.ClassName("android.widget.TextView"));
-      Assert.True(allProductsName.Count > 0);
-    }
-  }
-}
