--- android/appium_dotnet_driver_4_examples/android.sln
@@ -1,17 +1,0 @@
-
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2012
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "android", "android.csproj", "{41750684-85C3-4F6C-9B7A-F8141025E30A}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{41750684-85C3-4F6C-9B7A-F8141025E30A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{41750684-85C3-4F6C-9B7A-F8141025E30A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{41750684-85C3-4F6C-9B7A-F8141025E30A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{41750684-85C3-4F6C-9B7A-F8141025E30A}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
