--- android/appium_dotnet_driver_4_examples/first-test/FirstTest.cs
@@ -1,32 +1,0 @@
-﻿using System;
-using System.Threading;
-using System.Collections.ObjectModel;
-using NUnit.Framework;
-using OpenQA.Selenium;
-using OpenQA.Selenium.Appium;
-using OpenQA.Selenium.Appium.Android;
-using OpenQA.Selenium.Support.UI;
-
-namespace android.first
-{
-    [TestFixture("first","pixel-3")]
-    public class SingleTest : BrowserStackNUnitTest
-    {
-        public SingleTest(string profile, string device) : base(profile,device){}
-
-        [Test]
-        public void searchWikipedia()
-        {
-            WebDriverWait wait = new WebDriverWait(driver, TimeSpan.FromSeconds(30));
-            AndroidElement searchElement = (AndroidElement)wait.Until(SeleniumExtras.WaitHelpers.ExpectedConditions.ElementToBeClickable(MobileBy.AccessibilityId("Search Wikipedia")));
-            searchElement.Click();
-
-            AndroidElement insertTextElement = (AndroidElement)wait.Until(SeleniumExtras.WaitHelpers.ExpectedConditions.ElementToBeClickable(By.Id("org.wikipedia.alpha:id/search_src_text")));
-            insertTextElement.SendKeys("BrowserStack");
-            Thread.Sleep(5000);
-
-            ReadOnlyCollection<AndroidElement> allProductsName = driver.FindElements(By.ClassName("android.widget.TextView"));
-            Assert.True(allProductsName.Count > 0);
-        }
-    }
-}
