+++ ios/NUnit-Appium-App-BrowserStack/NUnit-Appium-App-BrowserStack.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 25.0.1705.1
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NUnit-Appium-App-BrowserStack", "NUnit-Appium-App-BrowserStack\NUnit-Appium-App-BrowserStack.csproj", "{ADCA9EA5-6CBF-4A4D-A504-EB9598646B9A}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{ADCA9EA5-6CBF-4A4D-A504-EB9598646B9A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{ADCA9EA5-6CBF-4A4D-A504-EB9598646B9A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{ADCA9EA5-6CBF-4A4D-A504-EB9598646B9A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{ADCA9EA5-6CBF-4A4D-A504-EB9598646B9A}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {5E596963-AA5A-4069-9ED2-EB29891EC158}
+	EndGlobalSection
+EndGlobal
