+++ ios/NUnit-Appium-App-BrowserStack/NUnit-Appium-App-BrowserStack/.config/dotnet-tools.json
@@ -1,0 +1,12 @@
+{
+  "version": 1,
+  "isRoot": true,
+  "tools": {
+    "browserstack-sdk": {
+      "version": "0.3.0",
+      "commands": [
+        "browserstack-sdk"
+      ]
+    }
+  }
+}
