--- ios/appium_dotnet_driver_3_examples/App.config
@@ -1,56 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<configuration>
-    <configSections>
-        <sectionGroup name="capabilities">
-            <section name="first" type="System.Configuration.AppSettingsSection, System.Configuration, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a"/>
-            <section name="local" type="System.Configuration.AppSettingsSection, System.Configuration, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a"/>
-            <section name="parallel" type="System.Configuration.AppSettingsSection, System.Configuration, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a"/>
-        </sectionGroup>
-        <sectionGroup name="environments">
-            <section name="iphone-11-pro" type="System.Configuration.AppSettingsSection, System.Configuration, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a"/>
-            <section name="iphone-11-pro-max" type="System.Configuration.AppSettingsSection, System.Configuration, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a"/>
-        </sectionGroup>
-    </configSections>
-
-    <appSettings>
-        <add key="user" value="<browserstack-user-name>"/>
-        <add key="key" value="<browserstack-access-key>"/>
-        <add key="server" value="hub.browserstack.com"/>
-    </appSettings>
-
-    <capabilities>
-        <first>
-            <add key="project" value="NUnit iOS"/>
-            <add key="build" value="First NUnit iOS Test"/>
-            <add key="name" value="first_test"/>
-            <add key="browserstack.debug" value="true"/>
-            <add key="app" value="bs://<app_hashed_id>"/>
-        </first>
-        <local>
-            <add key="project" value="NUnit iOS"/>
-            <add key="build" value="Local NUnit iOS Test"/>
-            <add key="name" value="local_test"/>
-            <add key="browserstack.debug" value="true"/>
-            <add key="browserstack.local" value="true"/>
-            <add key="app" value="bs://<app_hashed_id>"/>
-        </local>
-        <parallel>
-            <add key="project" value="NUnit iOS"/>
-            <add key="build" value="Parallel NUnit iOS Test"/>
-            <add key="name" value="parallel_test"/>
-            <add key="browserstack.debug" value="true"/>
-            <add key="app" value="bs://<app_hashed_id>"/>
-        </parallel>
-    </capabilities>
-
-    <environments>
-        <iphone-11-pro>
-            <add key="device" value="iPhone 11 Pro"/>
-            <add key="os_version" value="13.0"/>
-        </iphone-11-pro>
-        <iphone-11-pro-max>
-            <add key="device" value="iPhone 11 Pro Max"/>
-            <add key="os_version" value="13.0"/>
-        </iphone-11-pro-max>
-    </environments>
-</configuration>
