--- ios/appium_dotnet_driver_4_examples/ios.sln
@@ -1,17 +1,0 @@
-
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2012
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ios", "ios.csproj", "{94A16168-94B2-4047-866D-96611EBF5FA7}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{94A16168-94B2-4047-866D-96611EBF5FA7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{94A16168-94B2-4047-866D-96611EBF5FA7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{94A16168-94B2-4047-866D-96611EBF5FA7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{94A16168-94B2-4047-866D-96611EBF5FA7}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
