--- ios/appium_dotnet_driver_4_examples/packages.config
@@ -1,14 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Appium.WebDriver" version="4.2.1" targetFramework="net461" />
-  <package id="BrowserStackLocal" version="1.4.0" targetFramework="net461" />
-  <package id="Castle.Core" version="4.4.1" targetFramework="net461" />
-  <package id="DotNetSeleniumExtras.PageObjects" version="3.11.0" targetFramework="net461" />
-  <package id="DotNetSeleniumExtras.WaitHelpers" version="3.11.0" targetFramework="net461" />
-  <package id="Newtonsoft.Json" version="12.0.3" targetFramework="net461" />
-  <package id="NUnit" version="3.12.0" targetFramework="net461" />
-  <package id="NUnit3TestAdapter" version="3.17.0" targetFramework="net461" />
-  <package id="Selenium.Support" version="3.141.0" targetFramework="net461" />
-  <package id="Selenium.WebDriver" version="3.141.0" targetFramework="net461" />
-  <package id="System.Runtime.InteropServices.RuntimeInformation" version="4.3.0" targetFramework="net461" />
-</packages>
