--- ios/appium_dotnet_driver_4_examples/parallel-test/ParallelTest.cs
@@ -1,30 +1,0 @@
-﻿using System;
-using NUnit.Framework;
-using OpenQA.Selenium.Appium;
-using OpenQA.Selenium.Appium.iOS;
-using OpenQA.Selenium.Support.UI;
-
-namespace ios.parallel
-{
-    [TestFixture("parallel", "iphone-11-pro")]
-    [TestFixture("parallel", "iphone-11-pro-max")]
-    [Parallelizable(ParallelScope.Fixtures)]
-    public class ParallelTest : BrowserStackNUnitTest
-    {
-        public ParallelTest(string profile, string device) : base(profile, device) {}
-
-        [Test]
-        public void textVerificationTest()
-        {
-            WebDriverWait wait = new WebDriverWait(driver, TimeSpan.FromSeconds(30));
-            IOSElement textButton = (IOSElement) wait.Until(SeleniumExtras.WaitHelpers.ExpectedConditions.ElementToBeClickable(MobileBy.AccessibilityId("Text Button")));
-            textButton.Click();
-
-            IOSElement textInput = (IOSElement) wait.Until(SeleniumExtras.WaitHelpers.ExpectedConditions.ElementToBeClickable(MobileBy.AccessibilityId("Text Input")));
-            textInput.SendKeys("hello@browserstack.com"+"\n");
-
-            IOSElement textOutput = (IOSElement) wait.Until(SeleniumExtras.WaitHelpers.ExpectedConditions.ElementToBeClickable(MobileBy.AccessibilityId("Text Output")));
-            Assert.AreEqual(textOutput.Text,"hello@browserstack.com");
-        }
-    }
-}
